audio: Update tone duration support from uint16 to uint32
 - Previously, tone duration was limited to the maximum value of uint16.
 - PAL supports tone duration up to the maximum value of uint32.
 - Updated tone and DTMF duration to support uint32 maximum value (treated as infinite).

CRs-Fixed: 4426758

+++ simulation/libs/audio/ToneGeneratorStreamImpl.cpp
@@ -23,7 +23,7 @@ ToneGeneratorStreamImpl::~ToneGeneratorStreamImpl() {
  * Generate tone for the given duration at given frequency and gain.
  */
 telux::common::Status ToneGeneratorStreamImpl::playTone(std::vector<uint16_t> frequency,
-    uint16_t duration, uint16_t gain, telux::common::ResponseCallback callback) {
+    uint32_t duration, uint16_t gain, telux::common::ResponseCallback callback) {
 
     intptr_t cmdId;
     telux::common::Status status;
